Updating refresh procedure

+++ app/src/main/java/edu/utap/watchlist/api/Movie.kt
@@ -8,7 +8,7 @@ data class Movie (
     @SerializedName("backdrop_path")
     val backdropPath: String?,
     @SerializedName("budget")
-    val budget: Int,
+    val budget: Long,
     @SerializedName("id")
     val id: Int,
     @SerializedName("original_title")
@@ -26,7 +26,7 @@ data class Movie (
     @SerializedName("vote_average")
     val voteAverage: Float,
     @SerializedName("vote_count")
-    val voteCount: Int,
+    val voteCount: Long,
     @SerializedName("overview")
     val overview: String?,
     @SerializedName("status")
@@ -34,7 +34,7 @@ data class Movie (
     @SerializedName("tagline")
     val tagline: String?,
     @SerializedName("revenue")
-    val revenue: Int,
+    val revenue: Long,
     @SerializedName("original_language")
     val originalLanguage: String,
 )
